+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.1.4] - 2024-10-27
+### Changed
+- Relocated some results methods to a runtimes internal module
+
 ## [3.1.3] - 2024-10-26
 ### Added
 - Added passing `original_exception` to fault `!` calls